+++ core/model/application.go
@@ -76,17 +76,17 @@ func (p Permission) CheckAssigners(assignerPermissions []string) error {
 		return nil
 	}
 	if len(p.Assigners) == 0 {
-		return errors.Newf("not defined assigners for %s permission", p.Name)
+		return errors.ErrorData(logutils.StatusMissing, "assigners", &logutils.FieldArgs{"name": p.Name})
 	}
 
 	authorizedAssigners := p.Assigners
 	for _, authorizedAssigner := range authorizedAssigners {
-		if !authutils.ContainsString(assignerPermissions, authorizedAssigner) {
-			return errors.Newf("assigner %s is not satisfied", authorizedAssigner)
+		if authutils.ContainsString(assignerPermissions, authorizedAssigner) {
+			return nil
 		}
 	}
-	//all assigners are satisfied
-	return nil
+	//no assigners are satisfied
+	return errors.ErrorAction(logutils.ActionValidate, "assigner permissions", &logutils.FieldArgs{"name": p.Name})
 }
 
 func (p Permission) String() string {
@@ -131,10 +131,10 @@ func (c AppOrgRole) CheckAssigners(assignerPermissions []string) error {
 	for _, permission := range c.Permissions {
 		err := permission.CheckAssigners(assignerPermissions)
 		if err != nil {
-			return errors.Wrapf("error checking role permission assigners", err)
+			return errors.WrapErrorAction(logutils.ActionValidate, "role permission", &logutils.FieldArgs{"id": c.ID, "name": c.Name}, err)
 		}
 	}
-	//it satisfies all permissions
+	//all permissions may be assigned
 	return nil
 }
 
@@ -166,24 +166,20 @@ func (cg AppOrgGroup) CheckAssigners(assignerPermissions []string) error {
 	}
 
 	//check permission
-	if len(cg.Permissions) > 0 {
-		for _, permission := range cg.Permissions {
-			err := permission.CheckAssigners(assignerPermissions)
-			if err != nil {
-				return err
-			}
+	for _, permission := range cg.Permissions {
+		err := permission.CheckAssigners(assignerPermissions)
+		if err != nil {
+			return errors.WrapErrorAction(logutils.ActionValidate, "group permission", &logutils.FieldArgs{"id": cg.ID, "name": cg.Name}, err)
 		}
 	}
 	//check roles
-	if len(cg.Roles) > 0 {
-		for _, role := range cg.Roles {
-			err := role.CheckAssigners(assignerPermissions)
-			if err != nil {
-				return err
-			}
+	for _, role := range cg.Roles {
+		err := role.CheckAssigners(assignerPermissions)
+		if err != nil {
+			return errors.WrapErrorAction(logutils.ActionValidate, "group role", &logutils.FieldArgs{"id": cg.ID, "name": cg.Name}, err)
 		}
 	}
-	//all assigners are satisfied
+	//all permissions and roles may be assigned
 	return nil
 }